workers/scrapy/requirements.txt
scrapy>=2.11.0
scrapy-playwright>=0.0.34
openai>=1.12.0
python-dotenv>=1.0.0
supabase>=2.3.0
pydantic>=2.6.0
playwright>=1.41.0
